Made multi module project
Still need to figure out how to separate the two projects without
loading two times the `kotlin("multiplatform")` plugin.
