+++ .gitignore
@@ -25,3 +25,6 @@ buildProfile/
 
 # MacOS
 .DS_Store
+
+# Ignore Gradle build output directory
+build
