+++ js-app/src/jsMain/resources/graphql/Points.graphql
