+++ js-app/src/jsMain/resources/graphql/Test.graphql
