+++ js-app/src/jsMain/resources/graphql/schema.graphqls
