+++ js-app/src/jsMain/resources/index.html
@@ -5,7 +5,7 @@
 </head>
 <body>
 <script type="text/javascript" src="https://cdn.jsdelivr.net/gh/JetBrains/lets-plot@v3.2.0/js-package/distr/lets-plot.min.js"></script>
-<script src="graphql-plot.js"></script>
+<script src="js-app.js"></script>
 <div>
     <h2>Kotlin/JS Lets-Plot Sample</h2>
     <div id="content"></div>
