+++ server/src/jvmMain/kotlin/gql/plt/server/components/schema/models/Point.kt
@@ -1,0 +1,3 @@
+package gql.plt.server.components.schema.models
+
+data class Point(val x: Int, val y: Int)
