+++ server/src/jvmMain/resources/application.conf
@@ -1,0 +1,9 @@
+ktor {
+    deployment {
+        port = 8080
+    }
+
+    application {
+        modules = [ gql.plt.server.components.GraphQLModuleKt.graphQLModule ]
+    }
+}
