+++ settings.gradle.kts
@@ -1,3 +1,2 @@
-
 rootProject.name = "graphql-plot"
-
+include("js-app", "server")
