--- src/jvmMain/kotlin/me.stefanofuri.application/server/schema/models/Point.kt
@@ -1,3 +1,0 @@
-package me.stefanofuri.application.server.schema.models
-
-data class Point(val x: Int, val y: Int)
