--- src/jvmMain/resources/application.conf
@@ -1,9 +1,0 @@
-ktor {
-    deployment {
-        port = 8080
-    }
-
-    application {
-        modules = [ me.stefanofuri.application.server.GraphQLModuleKt.graphQLModule ]
-    }
-}
